Extract Store::{take_next_job, take_next_n_jobs} into impl module.
Following the pattern from 98c2cf8c8663cbf03a8c25b0662bc7ed2d3effe6 and
previous commits, this time for the job-acquisition related methods to
further break down the Store into smaller units of code.
